+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added a `Date` class.
 - Added a `Time` class.
 - Added a `remaining_days` property to the `Interval` class.
+- Added a `in_timestamp` property to the `Pendulum` class to retrieve the behavior of the now deprecated `timestamp` property.
 
 ### Changed
 
@@ -19,7 +20,6 @@
 ### Deprecated
 
 - `timestamp` should now be used as a method and no longer as a property. It will be a native method in the next major version.
-- `float_timestamp` is now deprecated. Use `timestamp()` instead.
 - `Interval` properties and methods related to years and months are now deprecated.
 - `Interval.days_exclude_weeks` is now deprecated. Use `remaining_days` instead.